Update patch set 2
Patch Set 2:

Let me clarify it.

SamsungServiceMode is a package that is not needed by all devices, hence a device that needs it needs to pull it manually by specifying this in the cm.dependencies file. Cid automatically chooses maguro for all translation changes. maguro does not need SamsungServiceMode, so the repository is not being downloaded and the cherry pick fails.

To solve this I have created a minimalistic device repository that contains a cm.dependencies file referring to SamsungServiceMode (and in jellybean to FMRadio). This just needs to be forked and set up in the build servers.

So, please verify this manually. No need to push additional patch sets for Cid. If the new device tree is used, I will announce this in the Google+ community and submit a new patch set myself.

Patch-set: 2